requirements.txt
pytest==6.2.3
selenium==3.14.0
allure-pytest==2.9.43
pytest-xdist==2.2.1
